+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Fix display issue with SIFTS mapping
+
 ## [v3.37.1] - 2023-06-20
 
 - Fix issues with wboit/dpoit in large scenes